Renaming netmhc_pan2.py to netmhc_pan28.py for consistency.
