+++ mhctools/__init__.py
@@ -13,7 +13,7 @@ from .netmhc3 import NetMHC3
 from .netmhc4 import NetMHC4
 from .netmhc_cons import NetMHCcons
 from .netmhc_pan import NetMHCpan
-from .netmhc_pan2 import NetMHCpan28
+from .netmhc_pan28 import NetMHCpan28
 from .netmhc_pan3 import NetMHCpan3
 from .netmhcii_pan import NetMHCIIpan
 from .random_predictor import RandomBindingPredictor
